feat: add code modification after review (https://github.com/watcha-fr/synapse/pull/113)
